timerevent, waitingroom, object/subject

+++ src/GameInterface/Type.h
@@ -19,7 +19,7 @@ using t_res = unsigned int;
 using t_cmd = unsigned int;
 using t_display = unsigned short;
 
-using t_id = uint16_t;
+using t_id = uint32_t;
 using t_id_atomic = std::atomic<t_id>;
 
 using t_time = std::time_t;
